+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RpcUrl.password` is now expected to be only string and default value is empty
   string. `RpcUrl.login_type` is now `PLAIN` to be valid for default password
   value.
+- `rpclogin` replaces `PLAIN` login with `SHA1` for increased security. The new
+  parameter `force_plain` was added to actually use `PLAIN` anyway.
 
 ### Fixed
 - Login to RPC Broker with PLAIN password when server has SHA1 configured for